1.0

+++ src/js/dean/delete/period.js
@@ -1,26 +1,29 @@
 function deletePeriod() {
-    $.ajax({
-        url: "modules/dean/delete/period.php",
-        method: "post",
-        data: {
-            id: currentPeriodId
-        },
-        success: function (data) {
-            if (data == 'ok') {
-                fetchAcademicPeriods().then(function () {
-                    fetchStudents().then(function () {
+    if (confirm("Are you sure to delete academic period?")) {
+        $.ajax({
+            url: app_url + "modules/dean/delete/period.php",
+            method: "post",
+            data: {
+                id: currentPeriodId
+            },
+            success: function (data) {
+                if (data == 'ok') {
+                    fetchAcademicPeriods().then(function () {
+                        fetchStudents().then(function () {
 
-                        alert("Period Deleted");
+                            alert("Period Deleted");
 
-                        deleteModal();
+                            deleteModal();
 
-                        studentsTab();
-                    })
-                });
+                            studentsTab();
+                        })
+                    });
+                }
+                else {
+                    alert("Failed to delete period")
+                }
             }
-            else {
-                alert("Failed to delete period")
-            }
-        }
-    })
+        })
+    }
+
 }
